pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "conciliador-bancario"
version = "0.1.0"
description = "Sistema de conciliación bancaria en Streamlit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.37.1",
    "pandas>=2.2.3",
    "openpyxl",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["logic*", "infra*"]
